+++ include/roq/oms/order.hpp
@@ -30,6 +30,7 @@ struct ROQ_PUBLIC Order final {
   double quantity = NaN;
   double price = NaN;
   double stop_price = NaN;
+  double risk = NaN;
   double remaining_quantity = NaN;
   double traded_quantity = NaN;
   double average_traded_price = NaN;
@@ -80,6 +81,7 @@ struct fmt::formatter<roq::oms::Order> {
         R"(quantity={}, )"
         R"(price={}, )"
         R"(stop_price={}, )"
+        R"(risk={}, )"
         R"(remaining_quantity={}, )"
         R"(traded_quantity={}, )"
         R"(average_traded_price={}, )"
@@ -114,6 +116,7 @@ struct fmt::formatter<roq::oms::Order> {
         value.quantity,
         value.price,
         value.stop_price,
+        value.risk,
         value.remaining_quantity,
         value.traded_quantity,
         value.average_traded_price,